Add icons to "Profile as" launch and context menus #19
All the "Run" and "Profile" related actions have been compared and the
missing icons were added to the "Profile as" actions.

The icon appears correctly in the "Menu visibility" configuration dialog
for the "Profile as" main menu group. The context menu has not been
tested due to not profiler available in my test runtime.