Have different overlays for launch icons in quick access #2
Use the icon associated with the launch mode as an overlay on top of the
icon registered for the launch type. That way, run/debug/profile of the
same launch can be distinguished.

The overlay icon in for default resolution is 12 pixels wide. The
typical overlay size of 7 pixels made it unrecognizable. The 2x
resolution overlay is 16 pixels wide.