properly resize Launch configuration controls
The text widget in both run and debug configurations follow
the sashing only partially - at some point, the filter text
overlays into the neighboring areas. this is because of the
conflicting layoutdata property defined for the text and the
label controls in the parent composite. if grabExcessHorizontalSpace
property is set to true for one and false for another child controls,
the layout gets convoluted, or follow only one of the request.

Fix is to force the same flag in both the cases leading to properly
laying out the children widgets in the composite.

Fixes: https://github.com/eclipse-platform/eclipse.platform.debug/issues/130