Launch shortcut to show action then configs
In some cases, the launch shortcut behavior is still desired and the
substitute launch config doesn't match this behavior (this might be
considered as a bug from the shortcut implementation, yet it's still an
important case that we can handle in the framework).
To avoid loosing functionality, we show the shortcut "as usual" and then
the launch configs that the shortcut may reference.

Fixes https://github.com/eclipse-platform/eclipse.platform.debug/issues/135