Hyperlinks are not accessible while the process is still logging #141
When a document change event is triggered, updating the links in the console at the
document change offset results in exiting the link which was being hovered before
the console's document changed (while the running process is still logging into the console).

Updating the links (triggering the linkEntered/linkExited methods) should be done only for the
current mouse offset when a document changes because it's the only link which might be caused
to be entered / exited after a document change.

Fixes https://github.com/eclipse-platform/eclipse.platform.debug/issues/141