Accidentally deleted all commits on my branch.  This commit restores the changes.  Plus it fixes unit test failures.
In the commits I lost, I was passing a parameter to the ProcessConole
constructor to enable command line history.  I determined this was not
the best way.  Now there is a method on IOConsolePage used to enable
command line history. Consider this scenario:
- User launches an application to be debugged.
- The launch delegate creates the process.
- The launch is added to the launch manager.
- ProcessConsoleManager listens for the new launch, creates
  a ProcessConsole, and adds it to the ConsoleManager.
  Notice the process console was created.  So there is no
  possibility of passing an enable command line history
  parameter to the constructor.
- The console view listens for ConsoleManager additions,
  which then creates a page for the console.  IConsolePageParticipants
  are called.
- The page participant is the place to enable command line history.

Created IIOConsolePage so command line history can be enabled.  The
history for a console has to enabled via the page so curosor management
can be performed.

The off by one error I mentioned earlier is not an error, the
original code was correct.

Add JUnit test for command line history.